requirements.txt
sqlalchemy
pymysql
python-dotenv
langchain
cryptography
streamlit
mysql-connector-python
llama-cpp-python
langchain_community
sqlglot